# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Board API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created user!"
                    },
                    "500": {
                        "description": "Error registration"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success login",
                        "schema": {
                            "properties": {
                                "access_token": {
                                    "type": "string"
                                },
                                "token_type": {
                                    "type": "string"
                                },
                                "expires_at": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Error login"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "500": {
                        "description": "LogOut error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of blog posts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagination number",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PostResource"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/user/{user}/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of user's blog posts",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PostResource"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/category/{category}/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of blog posts in certainly category",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PostResource"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/favorites/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of blog posts in my favorites",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PostResource"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/post/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get blog post by id",
                "description": "Get blog post by id",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/PostResource"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Post is not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/create/post": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "name": "image",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "You success create post"
                    },
                    "500": {
                        "description": "Error create post"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/delete/post/{post}": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "You success delete post"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "500": {
                        "description": "Error delete post"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/add/favor/{post}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Add a post to favorites",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "properties": {
                                "post": {
                                    "description": "post id",
                                    "type": "string"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Post already add to favorites or You can't add your post to favorites post"
                    },
                    "201": {
                        "description": "You success add post to favor"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/delete/favor/{post}": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post form favorites",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "You success delete post from favorites"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Post is not your"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "PostResource": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "category": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "image": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}